+++ server/plugins/Gotify.js
@@ -1,9 +1,10 @@
-var https = require('http');
+var http = require('http');
 var logger = require('../log');
 
 function run(trigger, scope, data, config, callback) {
     var pConf = data.pluginconf.Gotify;
     if (pConf && pConf.enable) {
+
         let port = 80; // default
         if (config.port) {
             port = parseInt(config.port);
@@ -14,12 +15,14 @@ function run(trigger, scope, data, config, callback) {
             priority = parseInt(pConf.priority);
         }
 
+        // POST data
         const dat = JSON.stringify({
             title: data.agency+' - '+data.alias,
             message: data.message,
             priority: priority
         });
 
+        // POST Options
         const options = {
             hostname: config.URL,
             port: port,
@@ -31,13 +34,15 @@ function run(trigger, scope, data, config, callback) {
             }
         };
 
-        const req = https.request(options, (res) => {
-            logger.main.debug('Gotify statusCode:' + res.statusCode);
+        // HTTP request
+        const req = http.request(options, (res) => {
+            logger.main.debug('Gotify: ' + res.statusCode);
             res.on('data', (d) => {
                 process.stdout.write(d)
             })
         });
 
+        // HTTP error
         req.on('error', (error) => {
             logger.main.error('Gotify:' + error);
         });